pyproject.toml
[project]
name = "bonk"
version = "1.0.0"
description = "Turning your PC into a 2004 DVD player on standby"
requires-python = ">=3.14"
license = { text = "MIT" }
dependencies = [
    "pygame-ce>=2.5.3",
]

[project.scripts]
bonk = "dvd_screensaver.__main__:main"

[project.optional-dependencies]
build = [
    "pyinstaller>=6.0",
]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dvd_screensaver"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
